Minesweeper Game
Creating a game using NumPy in which a random set of positive and negative numbers will be shuffled. Assigning all the negative numbers as bombs and taking the input from the user to select a row and column. Also asking the user for the total number of rows and columns to be assigned. Printing the raw format of the matrix box for reference.